# colblk: fix for BenchmarkUnsafeIntegerSlice